% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_INMB_DIRECT.R
\name{create_object_var_inmb_direct}
\alias{create_object_var_inmb_direct}
\title{Create an object var_inmb_direct}
\usage{
create_object_var_inmb_direct(var_inmb)
}
\arguments{
\item{var_inmb}{: variance of the Incremental Net Monetary Benefit}
}
\value{
create_object_var_inmb_direct returns an object of class \link{VAR_INMB_DIRECT}
}
\description{
When absolutely no data regarding the variability of costs and effectiveness are available, it is possible to directly provide a value for the variance of the Incremental Net Monetary Benefit in this object.
}
\examples{
## Create a var_inmb object :
object_var_inmb <- create_object_var_inmb_direct(var_inmb = 18324000)
## retrieve the inmb value from the object
object_var_inmb$get_var_inmb()
}
\seealso{
\link{create_object_var_inmb_diff} to calculate the theoretical standard deviation of the expected INB with different standard deviation in the reference and the experimental group

\link{create_object_var_inmb} to calculate the theoretical standard deviation of the expected INB with the same standard deviation in the reference and the experimental group
}
