% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_sujets.R
\name{gamma_risk}
\alias{gamma_risk}
\title{Function to estimate the gamma risk}
\usage{
gamma_risk(object_evpi_decrease)
}
\arguments{
\item{object_evpi_decrease}{An evpi_decrease object. See \link{create_object_evpi_decrease}}
}
\value{
A numeric value representing the gamma risk (probability)
}
\description{
The gamma risk is the probability that a decision based on the expected mean of the Incremental Net Monetary Benefit (INMB) is wrong.
In other terms, it is one minus the cost-effectiveness probability of an intervention.
Use \link{sample_size} function first to estimate the sample size before calling this function.
}
