% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INMB_DIRECT.R
\docType{class}
\name{INMB_DIRECT}
\alias{INMB_DIRECT}
\title{A Reference Class to represent the INMB (Incremental Net Monetary Benefit)}
\description{
If the INMB can be drawn from de, dc and lambda with \link{create_object_inmb}, one can also make directly an hypothesis on the value of the INMB.
}
\section{Fields}{

\describe{
\item{\code{inmb}}{: INMB expected Incremental Net monetary Benefit.}
}}


\section{Methods}{

 \describe{
 \item{get_inmb():}{Returns the Incremental Net Monetary Benefit (inmb)}
 \item{set_inmb(inmb):}{sets the inmb of this INMB_DIRECT object}
 }
}

\seealso{
\link{create_object_inmb_direct} the constructor

\link{create_object_inmb} to calculate the INMB
}
