% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lambda.R
\docType{class}
\name{Lambda}
\alias{Lambda}
\title{A Reference Class to represent the lambda value}
\description{
Lambda is known as the willingness to pay. That is the ceiling cost-effectiveness ratio or the maximum acceptable cost of a unit of effectiveness.
}
\section{Fields}{

\describe{
\item{\code{lambda}}{: Lambda is a monetary value. For example, the value of lambda is usually between 20 000 and 40 000 pounds/QALY in UK.}
}}


\section{Methods}{

 \describe{
 \item{set_lambda(lambda):}{sets the lambda value of this Lambda object}
 }
}

\seealso{
\link{create_object_lambda} the constructor
}
