% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INMB_DIRECT.R
\name{create_object_inmb_direct}
\alias{create_object_inmb_direct}
\title{Create an object INMB_DIRECT}
\usage{
create_object_inmb_direct(inmb)
}
\arguments{
\item{inmb}{: expected Incremental Net Monetary Benefit}
}
\value{
create_object_inmb_direct returns an object of class \link{INMB_DIRECT}
}
\description{
If the INMB can be drawn from de, dc and lambda with \link{create_object_inmb}, one can also make directly an hypothesis on the value of the INMB.
}
\examples{
## Create an object inmb_direct
object_inmb_direct <- create_object_inmb_direct (968)
}
\seealso{
\link{create_object_inmb}
}

