% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lambda.R
\name{create_object_lambda}
\alias{create_object_lambda}
\title{Create an object lambda}
\usage{
create_object_lambda(lambda)
}
\arguments{
\item{lambda}{: Lambda is a monetary value. For example, the value of lambda is usually between 20 000 and 40 000 pounds/QALY in UK.}
}
\value{
create_object_lambda returns an object of class \link{Lambda}
}
\description{
Lambda is known as the willingness to pay. That is the ceiling cost-effectiveness ratio or the maximum acceptable cost of a unit of effectiveness.
It must be coherent with the criteria of effectiveness used in the analysis (year of life, QALY, life saved, or a criteria related to morbidity).
}
\examples{
## Create an object lambda
object_lambda <- create_object_lambda (20000)
## retrieve the lambda value from the object
object_lambda$get_lambda()
}

