% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonctions_sujets.R
\name{sample_size}
\alias{sample_size}
\title{Function to calculate the estimated sample size based on the Expected Value of Perfect Information (EVPI)}
\usage{
sample_size(object_evpi_decrease, cost_indiv)
}
\arguments{
\item{object_evpi_decrease}{An evpi_decrease object. See \link{create_object_evpi_decrease}}

\item{cost_indiv}{cost of an additional inclusion in your planned cost-effectiveness study.}
}
\value{
A dataframe containing three columns :
   \item{N}{Total sample size of the planned cost-effectiveness study}
   \item{N_exp}{Sample size in the experimental group}
   \item{N_ref}{Sample size in the reference group}
}
\description{
This function calculates the optimal total sample size for a planned cost-effectiveness study.
The optimal sample size is reached when the marginal EVPI gain is less than or equal to
 the marginal cost of inclusion: (step_ref + step_exp) * cost_indiv.
}
