% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VAR_INMB_DIRECT.R
\docType{class}
\name{VAR_INMB_DIRECT}
\alias{VAR_INMB_DIRECT}
\title{A Reference Class to represent the theoretical standard deviation of the expected INB}
\description{
When absolutely no data regarding the variability of costs and effectiveness are available, it is possible to directly provide a value for the variance of the Incremental Net Monetary Benefit in this object.
}
\section{Fields}{

\describe{
\item{\code{var_inmb}}{: variance of the Incremental Net Monetary Benefit}
}}
\section{Methods}{

 \describe{
 \item{set_var_inmb(inmb):}{sets the var_inmb for this VAR_INMB_DIRECT object}
 }
}
\examples{
## Create a var_inmb object :
object_var_inmb <- create_object_var_inmb_direct(var_inmb = 18324000)
## retrieve the inmb value from the object
object_var_inmb$get_var_inmb()
}
\seealso{
\link{create_object_var_inmb_diff} to calculate the theoretical standard deviation of the expected INB with different standard deviation in the reference and the experimental group

\link{create_object_var_inmb} to calculate the theoretical standard deviation of the expected INB with the same standard deviation in the reference and the experimental group

\link{create_object_var_inmb_direct} the constructor
}

